+++ packages/frontend/CHANGELOG.md
@@ -1,5 +1,31 @@
 # graphql-portal-dashboard-frontend
 
+## 0.18.0
+
+### Minor Changes
+
+- 9807de8: auto login effect added
+
+### Patch Changes
+
+- bbac066: fix env vars
+
+## 0.17.0
+
+### Minor Changes
+
+- 35478c1: Fetcher changed
+
+### Patch Changes
+
+- 399c63b: deps update, sources without 2 step moved to consts
+
+## 0.16.1
+
+### Patch Changes
+
+- 706d4ca: fixed neo4j url validation and removed introspection
+
 ## 0.16.0
 
 ### Minor Changes